opt R.jar parse logic

+++ resguarder-plugin/src/main/kotlin/com/kernelflux/plugin/resguarder/ResguarderExtension.kt
@@ -3,7 +3,7 @@ package com.kernelflux.plugin.resguarder
 open class ResguarderExtension {
     var maxWidth: Int = 400
     var maxHeight: Int = 400
-    var maxFileSize: Int = 100 * 1024
+    var maxFileSize: Long = 100 * 1024
     var allBitmapUseImageLoader: Boolean = true
     var enableFileLog: Boolean = false
 }
